focused on input form
Work on making the input form area smaller and inline with the random and submit button.
Changed the random icon into a button with the font Awesome icon.

Need to put work into centering all the Search input area next time.
Need to adjust the appearance of bullets and text when searching for items.

Need to create function to search when the enter button is hit.